backend/requirements.txt
transformers
torch
mysql-connector-python
fastapi
uvicorn
groq
pymysql
python-dotenv